Add github repo embed. Moves gist logic to more specialised embed.
